Add a package_managers module
Before this commit every Importer had it's own VersionAPI class, which was
used to fetch all versions of a package from the corresponding package manager.

This was problematic due to 2 reasons (1)  Reusing these VersionAPI was hacky
(2) The importer files would contain most of the code related to fetching this
version info.

This is fixed by putting all this VersionAPI classes into one
vulnerabilities/package_managers.py module and the importers will now import
the classes from here.
