+++ vulnerabilities/importers/github.py
@@ -29,7 +29,6 @@ from typing import Tuple
 from typing import List
 from typing import Mapping
 from typing import Optional
-import xml.etree.ElementTree as ET
 
 import requests
 from dephell_specifier import RangeSpecifier
@@ -40,6 +39,9 @@ from vulnerabilities.data_source import DataSource
 from vulnerabilities.data_source import DataSourceConfiguration
 from vulnerabilities.data_source import Reference
 
+from vulnerabilities.package_managers import MavenVersionAPI
+from vulnerabilities.package_managers import NugetVersionAPI
+from vulnerabilities.package_managers import ComposerVersionAPI
 
 # set of all possible values of first '%s' = {'MAVEN','COMPOSER', 'NUGET'}
 # second '%s' is interesting, it will have the value '' for the first request,
@@ -213,116 +215,3 @@ class GitHubAPIDataSource(DataSource):
         affected_versions = {
             version for version in all_versions if version in version_range}
         return (affected_versions, all_versions - affected_versions)
-
-
-class MavenVersionAPI:
-    def __init__(self):
-        self.cache = {}
-
-    def get(self, pkg_name: str) -> Set[str]:
-        return self.cache.get(pkg_name, set())
-
-    def load_to_api(self, pkg_name: str) -> None:
-        if pkg_name in self.cache:
-            return
-
-        artifact_comps = pkg_name.split(":")
-        endpoint = self.artifact_url(artifact_comps)
-        resp = requests.get(endpoint).content
-
-        try:
-            xml_resp = ET.ElementTree(ET.fromstring(resp.decode("utf-8")))
-            self.cache[pkg_name] = self.extract_versions(xml_resp)
-        except ET.ParseError:
-            self.cache[pkg_name] = set()
-
-    @staticmethod
-    def artifact_url(artifact_comps: List[str]) -> str:
-        base_url = "https://repo.maven.apache.org/maven2/{}"
-        group_id, artifact_id = artifact_comps
-        group_url = group_id.replace(".", "/")
-        suffix = group_url + "/" + artifact_id + "/" + "maven-metadata.xml"
-        endpoint = base_url.format(suffix)
-
-        return endpoint
-
-    @staticmethod
-    def extract_versions(xml_response: ET.ElementTree) -> Set[str]:
-        all_versions = set()
-        for child in xml_response.getroot().iter():
-            if child.tag == "version":
-                all_versions.add(child.text)
-
-        return all_versions
-
-
-class NugetVersionAPI:
-    def __init__(self):
-        self.cache = {}
-
-    def get(self, pkg_name: str) -> Set[str]:
-        return self.cache.get(pkg_name.lower(), set())
-
-    def load_to_api(self, pkg_name: str) -> None:
-        if pkg_name in self.cache:
-            return
-        endpoint = self.nuget_url(pkg_name)
-        try:
-            resp = requests.get(endpoint).json()
-        # pkg_name=Microsoft.NETCore.UniversalWindowsPlatform triggers
-        # JSONDecodeError.
-        except json.decoder.JSONDecodeError:
-            self.cache[pkg_name.lower()] = set()
-            return
-
-        self.cache[pkg_name.lower()] = self.extract_versions(resp)
-
-    @staticmethod
-    def nuget_url(pkg_name: str) -> str:
-        base_url = "https://api.nuget.org/v3/registration5-semver1/{}/index.json"
-        return base_url.format(pkg_name.lower())
-
-    @staticmethod
-    def extract_versions(resp: dict) -> Set[str]:
-        all_versions = set()
-
-        try:
-            for entry in resp["items"][0]["items"]:
-                all_versions.add(entry["catalogEntry"]["version"])
-        # json response for YamlDotNet.Signed triggers this exception
-        except KeyError:
-            pass
-
-        return all_versions
-
-
-class ComposerVersionAPI:
-    def __init__(self):
-        self.cache = {}
-
-    def get(self, pkg_name: str) -> Set[str]:
-        return self.cache.get(pkg_name.lower(), set())
-
-    def load_to_api(self, pkg_name: str) -> None:
-        if pkg_name in self.cache:
-            return
-
-        endpoint = self.composer_url(pkg_name)
-        json_resp = requests.get(endpoint).json()
-        self.cache[pkg_name] = self.extract_versions(json_resp, pkg_name)
-
-    @staticmethod
-    def composer_url(pkg_name: str) -> str:
-        vendor, name = pkg_name.split("/")
-        return f"https://repo.packagist.org/p/{vendor}/{name}.json"
-
-    @staticmethod
-    def extract_versions(resp: dict, pkg_name: str) -> Set[str]:
-        all_versions = resp["packages"][pkg_name].keys()
-        # This filter ensures, that all_versions contains only released versions
-        all_versions = set(filter(lambda x: "dev" not in x, all_versions))
-        # See https://github.com/composer/composer/blob/44a4429978d1b3c6223277b875762b2930e83e8c/doc/articles/versions.md#tags  # nopep8
-        # for explanation of removing 'v'
-        all_versions = set(map(lambda x: x.replace("v", ""), all_versions))
-
-        return all_versions
